feat: Correctly substitute documentation links for router-based navigation
Some `href` attributes of the documentation links point to the static
FastAPI route, e.g. `http://localhost:3000/static/projects/name/version/`.
Since we want to use tanstack router for navigation, the links need to be
substituted with the UI navigation schema, in this case
`http://localhost:3000/name/version`.

This way, links from the documentation can be copied and shared.