fix: Do not return parsed `packaging.version.Version` objects but strings
Python's `packaging.version.Version` parses pre releases in a short-hand
notation, e.g. `2.0.0-beta` is parsed to `2.0.0-b0`. Since the directory
structure uses the actual git tag, there is a 404 error when requesting
`/static/projects/<name>/<parsed-tag>` since the directory name differs.

With this PR, the tags are still being validated to be SemVer, but the
returned version string is the directory name (and therefore the git tag).

This fixes #70.