refactor(api)!: Define order of included routers with a `FastAPI` lifespan
The loading order and definition of included routers in the API's
`__init__.py` file has grown quite complex. Since the order of loaded
routes is important for the API functionality and webapp integration,
it is crucial to make this more maintainable.

Moreover, with this change it is no longer necessary to monkeypatch
environment variables in tests for dynamic route/value patching
(e.g. for the app version).

This PR aims to refactor this process in a more transparent way using
FastAPI's [lifespan events](https://fastapi.tiangolo.com/advanced/events/#lifespan).

> [!WARNING]
> The API endpoint for fetching the version has changed from `/api/version` to `/api/version/` (trailing slash)
> This is the reason why this PR is marked as breaking change.